Move integration helpers into own file
